+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-04-22T07:28:50.306Z
+Generated: 2026-04-23T07:47:37.837Z
 
 - 2026-07-01 • Microsoft Sentinel • Action required • Call to action: update queries and automation by July 1, 2026 - standardized account entity naming in incidents and alerts
 - 2026-07-01 • Microsoft Sentinel • GA • , Microsoft Sentinel will be supported in the Defender portal only, and any remaining customers using the Azure portal will be automatically redirected.
@@ -11,7 +11,8 @@ Generated: 2026-04-22T07:28:50.306Z
 - 2026-05-01 • Microsoft Entra • Deprecation • Plan for change – Agent Registry consolidation into Microsoft Agent 365
 - 2026-04-01 • Microsoft Defender for Cloud • GA • General availability of container security capabilities in Azure Government cloud
 - 2026-04-01 • Microsoft Defender for Cloud • GA • General availability of Defender for Storage integration in Azure Portal Storage Center
-- 2026-04-01 • Microsoft Defender for Endpoint • Preview • Release - Windows
+- 2026-04-01 • Microsoft Defender for Endpoint • Preview • Release - macOS
+- 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126032.16.0 released: See
 - 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • enhancements and features for this release
 - 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126022.20.0 released: See
 - 2026-04-01 • Microsoft Defender for Endpoint • Ikke oppgitt • Release version 20.126022.18.0 released: See
@@ -21,4 +22,3 @@ Generated: 2026-04-22T07:28:50.306Z
 - 2026-04-01 • Microsoft Purview • GA • Data loss prevention policy tip reference for Outlook for Android, iOS, and macOS
 - 2026-04-01 • Microsoft Purview • Preview • DLP policies for unmanaged cloud apps support a new
 - 2026-04-01 • Microsoft Purview • Preview • Email notifications for browser and network DLP
-- 2026-04-01 • Microsoft Sentinel • GA • Microsoft Sentinel data federation (Preview)